requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-dotenv==1.0.0
pydantic==2.5.3
requests==2.31.0
tweepy==4.14.0
vaderSentiment==3.3.2
Pillow==10.2.0
schedule==1.2.1
slack-sdk==3.27.0
